+++ android/src/main/kotlin/me/yohom/foundation_fluttify/FoundationFluttifyPlugin.kt
@@ -11,11 +11,14 @@ import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.common.PluginRegistry.Registrar
 import me.yohom.foundation_fluttify.android.app.ActivityHandler
+import me.yohom.foundation_fluttify.android.app.ApplicationHandler
+import me.yohom.foundation_fluttify.android.app.NotificationHandler
 import me.yohom.foundation_fluttify.android.app.PendingIntentHandler
 import me.yohom.foundation_fluttify.android.content.IntentHandler
 import me.yohom.foundation_fluttify.android.graphics.BitmapHandler
 import me.yohom.foundation_fluttify.android.graphics.PointHandler
 import me.yohom.foundation_fluttify.android.location.LocationHandler
+import me.yohom.foundation_fluttify.android.os.BundleHandler
 import me.yohom.foundation_fluttify.android.util.PairHandler
 
 // The stack that exists on the Dart side for a method call is enabled only when the MethodChannel passing parameters are limited
@@ -53,8 +56,11 @@ class FoundationFluttifyPlugin : FlutterPlugin, ActivityAware, MethodCallHandler
         val args = methodCall.arguments as? Map<String, Any> ?: mapOf()
         methodCall.method.run {
             when {
-                startsWith("android.app.Activity") -> ActivityHandler(methodCall.method, args, methodResult)
+                startsWith("android.app.Application") -> ApplicationHandler(methodCall.method, args, methodResult, activity)
+                startsWith("android.app.Activity") -> ActivityHandler(methodCall.method, args, methodResult, activity)
                 startsWith("android.app.PendingIntent") -> PendingIntentHandler(methodCall.method, args, methodResult)
+                startsWith("android.app.Notification") -> NotificationHandler(methodCall.method, args, methodResult, activity)
+                startsWith("android.os.Bundle") -> BundleHandler(methodCall.method, args, methodResult)
                 startsWith("android.content.Intent") -> IntentHandler(methodCall.method, args, methodResult)
                 startsWith("android.graphics.Bitmap") -> BitmapHandler(methodCall.method, args, methodResult)
                 startsWith("android.graphics.Point") -> PointHandler(methodCall.method, args, methodResult)