+++ example/lib/main.dart
@@ -1,6 +1,10 @@
 import 'package:decorated_flutter/decorated_flutter.dart';
 import 'package:flutter/material.dart';
 import 'package:foundation_fluttify/foundation_fluttify.dart';
+import 'package:foundation_fluttify_example/src/demo/type/core/core.screen.dart';
+import 'package:foundation_fluttify_example/src/demo/type/platform/platform.screen.dart';
+import 'package:foundation_fluttify_example/src/widget/function_group.widget.dart';
+import 'package:foundation_fluttify_example/src/widget/function_item.widget.dart';
 
 void main() => runApp(MyApp());
 
@@ -16,10 +20,30 @@ class _MyAppState extends State<MyApp> {
       home: Scaffold(
         appBar: AppBar(title: const Text('Plugin example app')),
         body: DecoratedColumn(
+          itemSpacing: kSpaceBig,
           children: <Widget>[
-            RaisedButton(
-              onPressed: _testAddProperty,
-              child: Text('AddProperty'),
+            FunctionGroup(
+              headLabel: 'Type Demo',
+              children: <Widget>[
+                FunctionItem(
+                  label: 'Core Type Demo',
+                  sublabel: 'CoreScreen',
+                  target: CoreScreen(),
+                ),
+                FunctionItem(
+                  label: 'Platform Type Demo',
+                  sublabel: 'PlatformScreen',
+                  target: PlatformScreen(),
+                ),
+              ],
+            ),
+            FunctionGroup(
+              headLabel: 'Object Demo',
+              children: <Widget>[],
+            ),
+            FunctionGroup(
+              headLabel: 'Function Demo',
+              children: <Widget>[],
             ),
           ],
         ),