+++ example/lib/src/demo/type/platform/ios_type/ui_view/rotate.screen.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+class UIViewRotateScreen extends StatefulWidget {
+  @override
+  _UIViewRotateScreenState createState() => _UIViewRotateScreenState();
+}
+
+class _UIViewRotateScreenState extends State<UIViewRotateScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text('UIViewRotate')),
+      body: Center(child: Text('UIViewRotate')),
+    );
+  }
+}