+++ ios/Classes/UIColorHandler.h
@@ -1,0 +1,15 @@
+//
+//  UIColorHandler.h
+//  foundation_fluttify
+//
+//  Created by Yohom Bao on 2020/3/9.
+//
+
+#import <Foundation/Foundation.h>
+#import <Flutter/Flutter.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+void UIColorHandler(NSString* method, NSDictionary* args, FlutterResult methodResult);
+
+NS_ASSUME_NONNULL_END