+++ lib/src/function/functions.dart
@@ -18,36 +18,24 @@ Future<T> platform<T>({
   _FutureCallback<T> android,
   _FutureCallback<T> ios,
 }) async {
-  if (android != null && Platform.isAndroid) {
-    final releasePool = <Ref>{};
-    T result;
-    try {
-      result = await android(releasePool);
-    } catch (e) {
-      return Future.error(e);
-    } finally {
+  final releasePool = <Ref>{};
+  try {
+    if (android != null && Platform.isAndroid) {
+      return await android(releasePool);
+    } else if (ios != null && Platform.isIOS) {
+      return await ios(releasePool);
+    } else {
+      return Future.value();
+    }
+  } catch (e) {
+    return Future.error(e);
+  } finally {
+    if (releasePool.isNotEmpty) {
       await releasePool.release_batch();
       releasePool.clear();
       // remove all local object from global object pool
       kNativeObjectPool.removeAll(releasePool);
     }
-    return result;
-  } else if (ios != null && Platform.isIOS) {
-    final releasePool = <Ref>{};
-    T result;
-    try {
-      result = await ios(releasePool);
-    } catch (e) {
-      return Future.error(e);
-    } finally {
-      await releasePool.release_batch();
-      releasePool.clear();
-      // remove all local object from global object pool
-      kNativeObjectPool.removeAll(releasePool);
-    }
-    return result;
-  } else {
-    return Future.value();
   }
 }
 
@@ -56,13 +44,17 @@ Future<void> clearHeap() async {
 }
 
 Future<void> pushStack(String name, Ref ref) async {
-  await kMethodChannel.invokeMethod(
-      'PlatformService::pushStack', {'name': name, 'refId': ref.refId});
+  await kMethodChannel.invokeMethod('PlatformService::pushStack', {
+    'name': name,
+    'refId': ref.refId,
+  });
 }
 
 Future<void> pushStackJsonable(String name, dynamic jsonable) async {
-  await kMethodChannel.invokeMethod(
-      'PlatformService::pushStackJsonable', {'name': name, 'data': jsonable});
+  await kMethodChannel.invokeMethod('PlatformService::pushStackJsonable', {
+    'name': name,
+    'data': jsonable,
+  });
 }
 
 Future<void> clearStack() async {