+++ lib/src/type/core/typedefs.dart
@@ -1,5 +1,8 @@
 import 'dart:async';
 
+import 'package:foundation_fluttify/foundation_fluttify.dart';
+import 'package:foundation_fluttify/src/type/platform/ios_type/ui_application.dart';
+
 typedef Future<void> FutureVoidCallback();
 typedef Future<void> FutureValueChanged<T>(T value);
 typedef FutureOr<void> Action0();
@@ -111,4 +114,60 @@ typedef FutureOr<R> Function9<T1, T2, T3, T4, T5, T6, T7, T8, T9, R>(
   T7 arg7,
   T8 arg8,
   T9 arg9,
+);
+
+typedef void ApplicationDidFinishLaunchingWithOptions(
+  UIApplication application,
+  Map options,
+);
+typedef void ApplicationWillFinishLaunchingWithOptions(
+  UIApplication application,
+  Map options,
+);
+typedef void ApplicationDidBecomeActive(UIApplication application);
+typedef void ApplicationWillResignActive(UIApplication application);
+typedef void ApplicationDidEnterBackground(UIApplication application);
+typedef void ApplicationWillEnterForeground(UIApplication application);
+typedef void ApplicationWillTerminate(UIApplication application);
+typedef void ApplicationDidRegisterUserNotificationSettings(
+  UIApplication application,
+  UIUserNotificationSettings notificationSettings,
+);
+typedef void ApplicationDidRegisterForRemoteNotificationsWithDeviceToken(
+  UIApplication application,
+  NSData deviceToken,
+);
+typedef void ApplicationDidReceiveRemoteNotificationFetchCompletionHandler(
+  UIApplication application,
+  Map userInfo,
+);
+typedef void ApplicationOpenURLOptions(
+  UIApplication application,
+  NSUrl url,
+  Map<String, dynamic> options,
+);
+typedef void ApplicationHandleOpenURL(
+  UIApplication application,
+  NSUrl url,
+);
+typedef void ApplicationOpenURLSourceApplicationAnnotation(
+  UIApplication application,
+  NSUrl url,
+  String sourceApplication,
+  NSObject annotation,
+);
+typedef void ApplicationPerformActionForShortcutItemCompletionHandler(
+  UIApplication application,
+  UIApplicationShortcutItem shortcutItem,
+);
+typedef void ApplicationHandleEventsForBackgroundURLSessionCompletionHandler(
+  UIApplication application,
+  String identifier,
+);
+typedef void ApplicationPerformFetchWithCompletionHandler(
+  UIApplication application,
+);
+typedef void ApplicationContinueUserActivityRestorationHandler(
+  UIApplication application,
+  NSUserActivity userActivity,
 );