+++ lib/src/type/platform/android_type/android/os/bundle.dart
@@ -4,8 +4,8 @@ import 'package:foundation_fluttify/src/type/platform/android_type/java/lang/obj
 // ignore_for_file: non_constant_identifier_names, camel_case_types, missing_return, unused_import
 class android_os_Bundle extends java_lang_Object {
   static Future<android_os_Bundle> create() async {
-    final refId = await kMethodChannel
-        .invokeMethod('PlatformService::createandroid_os_Bundle');
+    final refId =
+        await kMethodChannel.invokeMethod('android.os.Bundle::create');
     return android_os_Bundle()
       ..refId = refId
       ..tag = 'platform';