+++ lib/src/type/platform/ios_type/ui_color.dart
@@ -6,8 +6,8 @@ import 'ns_object.dart';
 
 class UIColor extends NSObject {
   static Future<UIColor> create(Color color) async {
-    final refId = await kMethodChannel.invokeMethod(
-        'PlatformService::createUIColor', {'colorValue': color.value});
+    final refId = await kMethodChannel
+        .invokeMethod('UIColor::create', {'colorValue': color.value});
     return UIColor()
       ..refId = refId
       ..tag = 'platform';