+++ lib/src/type/platform/ios_type/ui_view.dart
@@ -4,6 +4,19 @@ import 'cg_rect.dart';
 import 'ns_object.dart';
 
 class UIView extends NSObject {
+  static Future<UIView> create() async {
+    final result = await kMethodChannel.invokeMethod('UIView::create');
+    return UIView()..refId = result;
+  }
+
+  Future<void> rotate(double angle) async {
+    assert(angle != null);
+    await kMethodChannel.invokeMethod('UIView::rotate', {
+      'refId': refId,
+      'angle': angle,
+    });
+  }
+
   Future<CGRect> get frame async {
     final result =
         await kMethodChannel.invokeMethod('UIView::getFrame', {'refId': refId});